Buscar provedores

+++ src/main/java/com/negocio/automoviles/models/Detalle.java
@@ -5,8 +5,8 @@ public class Detalle {
     private int ProveedorID;
     private int ParteID;
     private int ConsecutivoOrden;
-    private float precio;
-    private float porcentaje_ganancia;
+    private double precio;
+    private double porcentaje_ganancia;
     private String nombre_parte;
     private String nombre_provedor;
 
@@ -18,13 +18,13 @@ public class Detalle {
 
     public void setNombre_parte(String nombre_parte) { this.nombre_parte = nombre_parte; }
 
-    public float getPorcentaje_ganancia() { return porcentaje_ganancia; }
+    public double getPorcentaje_ganancia() { return porcentaje_ganancia; }
 
-    public void setPorcentaje_ganancia(float porcentaje_ganancia) { this.porcentaje_ganancia = porcentaje_ganancia; }
+    public void setPorcentaje_ganancia(double porcentaje_ganancia) { this.porcentaje_ganancia = porcentaje_ganancia; }
 
-    public float getPrecio() { return precio; }
+    public double getPrecio() { return precio; }
 
-    public void setPrecio(float precio) { this.precio = precio; }
+    public void setPrecio(double precio) { this.precio = precio; }
 
     public int getCantidad() {
         return cantidad;
